Import numpy eagerly on Windows
The numpy library has a bug when importing on Windows with stdin already
open, which causes numpy to hang, because it also tries to read from
stdin internally. See: https://github.com/numpy/numpy/issues/24290

To avoid this, a proper fix will be to add a feature to Appose to pass
some kind of init script that is run early, before the Appose pipes get
established. Then it would be as simple as passing a single line
`import numpy` as the init script in code that needs to avoid this bug.

But because SAMJ currently ships its own zipped copy of appose-python,
we add the following block of code to appose-python's python_worker.py:

    # HACK: Avoid numpy hang on Windows systems when stdin is open.
    # https://github.com/numpy/numpy/issues/24290
    if sys.platform.startswith("win"):
        try:
            import numpy
        except Exception as e:
            print(f"[WARNING] Failed to import numpy eagerly: {e}", file=sys.stderr)

(You can't see the patch directly here since it's inside the zip file.)